Convert nativerw-sidekick@.service to use VCB
We follow the new simpler convention for setting service information.
This means we now depend on vulcan config builder to set up vulcan for
us.

Note that we still set the health check info directly for the moment,
until aggregate health check is later updated to query the new standard.